+++ CHANGELOG.md
@@ -24,6 +24,10 @@ Nothing yet.
   - Chart area no longer overflows below X or Y axes when setting custom scale
     domains
   - Band axis ticks should now be aligned correctly
+  - Support of urn CRS in WMTS / WMS layers (enabled are ESPG:3857)
+  - Support for getTile operation on WMTS / WMS
+  - Update providers list according to
+    https://rdataflow.github.io/WMTS-WMS-monitor
 
 # 5.7.6 - 2025-05-06